Realizadas las correcciones comentadas en el PR
