+++ project-addons/customer_area/res_partner_view.xml
@@ -19,10 +19,22 @@
             <field name="inherit_id" ref="base.view_partner_form" />
             <field name="arch" type="xml">
                 <field name="website" position="before">
-                    <field name="area_id" attrs="{'invisible':[('customer', '=', False)]}"/>
+                    <field name="area_id" on_change="change_sales_team(area_id)" attrs="{'required':True, 'invisible':[('customer', '=', False)]}"/>
                 </field>
             </field>
         </record>
+
+        <!--<record id="view_partner_form_add_require_area" model="ir.ui.view">
+            <field name="name">res.partner.form.add_require_area</field>
+            <field name="model">res.partner</field>
+            <field name="inherit_id" ref="customer_area.res_partner_view_area_form"/>
+            <field name="arch" type="xml">
+                <field name="area_id" position="attributes">
+                    <attribute name="on_change"></attribute>
+                    <attribute name="required">True</attribute>
+                </field>
+            </field>
+        </record>-->
 
     </data>
 </openerp>
