+++ project-addons/product_outlet/wizard/product_outlet_wizard.py
@@ -21,6 +21,22 @@
 
 from openerp import models, fields, api, exceptions, _
 
+class ProductCategory(models.Model):
+
+    _inherit = 'product.category'
+
+    percent = fields.Float(string="Percent",
+                           help="This percent will be used when a product moves to an outlet category")
+
+    @api.one
+    @api.constrains('percent')
+    def check_length(self):
+        percent = self.percent
+        if (percent > 100) | (percent < 0):
+            raise exceptions. \
+                ValidationError(_('Error ! The percent values must be between 0 and 100'))
+        return True
+
 class product_outlet_wizard(models.TransientModel):
     _name = "product.outlet.wizard"
