+++ project-addons/product_outlet/wizard/product_outlet_wizard_view.xml
@@ -39,6 +39,18 @@
         view_mode="form"
         target="new"/>
 
+        <record id="product_category_add_percent_form" model="ir.ui.view">
+            <field name="name">product.category.form.percent</field>
+            <field name="model">product.category</field>
+            <field name="inherit_id" ref="product.product_category_form_view"/>
+            <field name="arch" type="xml">
+                <field name="type" position="after">
+                    <field name="percent"/>
+                </field>
+            </field>
+        </record>
+
+
     </data>
 </openerp>
